FastSim scripts now write to /data/schoef to avoid lost tokens

+++ MonoJetAnalysis/crab/runFastSim.py
@@ -18,14 +18,16 @@ while True:
   if stop - skipEvents<=0: break
 
   maxEvents = min(maxJobLength, stop - skipEvents)
-  
-  
-  sstring = "mkdir "+uniqueDirname+";cd "+uniqueDirname+";"
 
-  sstring+="cmsRun /afs/hephy.at/scratch/s/schoefbeck/CMS/CMSSW_5_3_11_patch6/src/Workspace/MonoJetAnalysis/crab/run_Hadronizer_SMS_Scans_2jets_Qcut44_TuneZ2star_8TeV_madgraph_tauola_cff_py_GEN_FASTSIM_HLT_PU.py"
-  sstring+=" infile=file:"+ifile+" maxEvents="+str(maxEvents)+" skipEvents="+str(skipEvents)+";"
-  sstring+="cd ..; rm -rf "+uniqueDirname
-  print sstring
-  os.system(sstring)
+  outFile = "/data/schoef/root/from_"+str(skipEvents)+"_to_"+str(skipEvents+maxEvents)+"_"+(ifile.split("/")[-1]).replace('.lhe', '.root').replace('file:', '') 
+  if not os.path.isfile(outFile):  
+    sstring = "cd /data/schoef/tmp; mkdir "+uniqueDirname+";cd "+uniqueDirname+";"
+    sstring+="cmsRun /afs/hephy.at/scratch/s/schoefbeck/CMS/CMSSW_5_3_11_patch6/src/Workspace/MonoJetAnalysis/crab/run_Hadronizer_SMS_Scans_2jets_Qcut44_TuneZ2star_8TeV_madgraph_tauola_cff_py_GEN_FASTSIM_HLT_PU.py"
+    sstring+=" infile=file:"+ifile+" outfile="+outFile+" maxEvents="+str(maxEvents)+" skipEvents="+str(skipEvents)+";"
+    sstring+="cd ..; rm -rf "+uniqueDirname
+    print sstring
+    os.system(sstring)
+  else:
+    print "Found",outFile,"->skipping!"
   skipEvents+=maxEvents
